used dedicated files for each template
