+++ canarytokens/windows_fake_fs_templates/folder_home_network.py
@@ -1,0 +1,44 @@
+HOME_NETWORK_STRUCTURE = [
+    {
+        "name": "HomeNetwork",
+        "type": "folder",
+        "children": [
+            {
+                "name": "NetworkDevices",
+                "type": "folder",
+                "children": [
+                    {"name": "RouterConfig.pdf", "type": "pdf"},
+                    {"name": "SwitchConfig.docx", "type": "docx"},
+                    {"name": "AccessPoints.xlsx", "type": "xls"},
+                ],
+            },
+            {
+                "name": "ISP",
+                "type": "folder",
+                "children": [
+                    {"name": "ISP_Contract.pdf", "type": "pdf"},
+                    {"name": "BillingStatements.xlsx", "type": "xls"},
+                    {"name": "SupportContacts.docx", "type": "docx"},
+                ],
+            },
+            {
+                "name": "NetworkDiagram",
+                "type": "folder",
+                "children": [
+                    {"name": "TopologyDiagram.pdf", "type": "pdf"},
+                    {"name": "IPAddresses.xlsx", "type": "xls"},
+                    {"name": "PortMapping.docx", "type": "docx"},
+                ],
+            },
+            {
+                "name": "Security",
+                "type": "folder",
+                "children": [
+                    {"name": "FirewallRules.docx", "type": "docx"},
+                    {"name": "DevicePasswords.xlsx", "type": "xls"},
+                    {"name": "SecurityChecklist.pdf", "type": "pdf"},
+                ],
+            },
+        ],
+    }
+]
