+++ canarytokens/windows_fake_fs_templates/folder_personal_correspondence.py
@@ -1,0 +1,44 @@
+PERSONAL_CORRESPONDENCE_STRUCTURE = [
+    {
+        "name": "PersonalCorrespondence",
+        "type": "folder",
+        "children": [
+            {
+                "name": "Family",
+                "type": "folder",
+                "children": [
+                    {"name": "Letters.docx", "type": "docx"},
+                    {"name": "Photos.pdf", "type": "pdf"},
+                    {"name": "Birthdays.xlsx", "type": "xls"},
+                ],
+            },
+            {
+                "name": "Friends",
+                "type": "folder",
+                "children": [
+                    {"name": "Emails.docx", "type": "docx"},
+                    {"name": "GroupPhotos.pdf", "type": "pdf"},
+                    {"name": "EventInvites.xlsx", "type": "xls"},
+                ],
+            },
+            {
+                "name": "WorkAcquaintances",
+                "type": "folder",
+                "children": [
+                    {"name": "LinkedInContacts.xlsx", "type": "xls"},
+                    {"name": "ThankYouNotes.docx", "type": "docx"},
+                    {"name": "NetworkingEvents.pdf", "type": "pdf"},
+                ],
+            },
+            {
+                "name": "Miscellaneous",
+                "type": "folder",
+                "children": [
+                    {"name": "HolidayCards.pdf", "type": "pdf"},
+                    {"name": "PenPals.docx", "type": "docx"},
+                    {"name": "CorrespondenceTracker.xlsx", "type": "xls"},
+                ],
+            },
+        ],
+    }
+]
