+++ canarytokens/windows_fake_fs_templates/folder_personal_finances.py
@@ -1,0 +1,77 @@
+PERSONAL_FINANCES_STRUCTURE = [
+    {
+        "name": "PersonalFinances",
+        "type": "folder",
+        "children": [
+            {
+                "name": "BankAccounts",
+                "type": "folder",
+                "children": [
+                    {"name": "Checking.xlsx", "type": "xls"},
+                    {"name": "Savings.xlsx", "type": "xls"},
+                    {"name": "CreditCards.xlsx", "type": "xls"},
+                ],
+            },
+            {
+                "name": "Investments",
+                "type": "folder",
+                "children": [
+                    {"name": "Stocks.xlsx", "type": "xls"},
+                    {"name": "Retirement.xlsx", "type": "xls"},
+                    {"name": "Crypto.xlsx", "type": "xls"},
+                ],
+            },
+            {
+                "name": "Budget",
+                "type": "folder",
+                "children": [
+                    {"name": "MonthlyBudget.xlsx", "type": "xls"},
+                    {"name": "ExpenseTracker.xlsx", "type": "xls"},
+                    {"name": "SavingsPlan.docx", "type": "docx"},
+                ],
+            },
+            {
+                "name": "Loans",
+                "type": "folder",
+                "children": [
+                    {"name": "Mortgage.pdf", "type": "pdf"},
+                    {"name": "CarLoan.pdf", "type": "pdf"},
+                    {"name": "StudentLoan.docx", "type": "docx"},
+                ],
+            },
+        ],
+    },
+    {
+        "name": "Taxes",
+        "type": "folder",
+        "children": [
+            {
+                "name": "Income",
+                "type": "folder",
+                "children": [
+                    {"name": "W2_Forms.pdf", "type": "pdf"},
+                    {"name": "1099_Forms.pdf", "type": "pdf"},
+                    {"name": "PayStubs.pdf", "type": "pdf"},
+                ],
+            },
+            {
+                "name": "Deductions",
+                "type": "folder",
+                "children": [
+                    {"name": "CharityReceipts.pdf", "type": "pdf"},
+                    {"name": "MedicalExpenses.docx", "type": "docx"},
+                    {"name": "EducationCosts.pdf", "type": "pdf"},
+                ],
+            },
+            {
+                "name": "Returns",
+                "type": "folder",
+                "children": [
+                    {"name": "LatestTaxReturn.pdf", "type": "pdf"},
+                    {"name": "PreviousTaxReturn.docx", "type": "docx"},
+                    {"name": "ArchivedTaxReturn.pdf", "type": "pdf"},
+                ],
+            },
+        ],
+    },
+]
