+++ canarytokens/windows_fake_fs_templates/folder_photo_archive.py
@@ -1,0 +1,96 @@
+PHOTO_ARCHIVE_STRUCTURE = [
+    {
+        "name": "PhotoArchive",
+        "type": "folder",
+        "children": [
+            {
+                "name": "Family",
+                "type": "folder",
+                "children": [
+                    {
+                        "name": "Vacations",
+                        "type": "folder",
+                        "children": [
+                            {"name": "BeachTrip.jpg", "type": "jpg"},
+                            {"name": "MountainHike.jpg", "type": "jpg"},
+                        ],
+                    },
+                    {
+                        "name": "Reunions",
+                        "type": "folder",
+                        "children": [
+                            {"name": "ReunionDinner.jpg", "type": "jpg"},
+                            {"name": "CousinsGathering.jpg", "type": "jpg"},
+                        ],
+                    },
+                ],
+            },
+            {
+                "name": "Friends",
+                "type": "folder",
+                "children": [
+                    {
+                        "name": "Parties",
+                        "type": "folder",
+                        "children": [
+                            {"name": "NewYearParty.jpg", "type": "jpg"},
+                            {"name": "BirthdayBash.jpg", "type": "jpg"},
+                        ],
+                    },
+                    {
+                        "name": "Trips",
+                        "type": "folder",
+                        "children": [
+                            {"name": "RoadTrip.jpg", "type": "jpg"},
+                            {"name": "CityExploration.jpg", "type": "jpg"},
+                        ],
+                    },
+                ],
+            },
+            {
+                "name": "Nature",
+                "type": "folder",
+                "children": [
+                    {
+                        "name": "Landscapes",
+                        "type": "folder",
+                        "children": [
+                            {"name": "SunsetOverHills.jpg", "type": "jpg"},
+                            {"name": "LakeReflection.jpg", "type": "jpg"},
+                        ],
+                    },
+                    {
+                        "name": "Wildlife",
+                        "type": "folder",
+                        "children": [
+                            {"name": "DeerInForest.jpg", "type": "jpg"},
+                            {"name": "BirdInFlight.jpg", "type": "jpg"},
+                        ],
+                    },
+                ],
+            },
+            {
+                "name": "Miscellaneous",
+                "type": "folder",
+                "children": [
+                    {
+                        "name": "RandomShots",
+                        "type": "folder",
+                        "children": [
+                            {"name": "StreetArt.jpg", "type": "jpg"},
+                            {"name": "AbstractPattern.jpg", "type": "jpg"},
+                        ],
+                    },
+                    {
+                        "name": "Projects",
+                        "type": "folder",
+                        "children": [
+                            {"name": "PhotoEssay1.jpg", "type": "jpg"},
+                            {"name": "PhotoEssay2.jpg", "type": "jpg"},
+                        ],
+                    },
+                ],
+            },
+        ],
+    }
+]
